requirements.txt
python-telegram-bot
python-dotenv~=0.12.0
requests~=2.23.0